Remove the thread optimisations
This seemed to cause us a bug, which we isolated to the thread local
and the way that Objectify works.
You could recreate by having a simple datastore counter, and having
two threads that attempt to increment it.
Thread A fetches the data, updates the counter and saves it.
Thread B fetches the data, updates the counter and saves it.
Thread A doesn't fetch the data, but uses a threadlocal cache,
updates the counter and saves it, overwriting Thread B's change.

It appears that Objectify.factory.begin has a local cache, so we
need to repeat it in between requests.  We decided for our app that
the cost of initialising Objectify per request was acceptable.